requirements.txt
pytorch>=2.0.0
torch-geometric>=2.3.0
pytorch-lightning>=2.0.0
kafka-python>=2.0.2
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.10.7
python-dotenv>=1.0.0
prometheus-client>=0.16.0
kubernetes>=26.1.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
